Updates across the board, please do not skip.
This update includes complete changes to the internal scattering code. The monte carlo methods remain the same, but no cloud boundaries are considered. The only boundaries that end the simulation are the provided COORDS - the simulation edges.
Thus for each scattering the path between the initial point and the end point of that scattering event is evaluated for changes in the mean free path. If all or none of the path is within the allowed threshold, then the end point is kept. If there is any change along the path the photon end position is updated to that point.

Additionally, debugging is helped in the ipynb version of the 777 and 337 code to output the photon path. Be careful that this is not kept on for many photons.